BACKEND: DB: users 'crud' added. 42 API getted data saving DB in process...

+++ srcs/requirements/backend/nest-js/src/users/users.controller.ts
@@ -1,34 +1,81 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, NotFoundException, HttpException, HttpStatus } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { isEqual } from 'lodash';
 
 @Controller('users')
 export class UsersController {
 	constructor(private readonly usersService: UsersService) {}
 
+	/**
+	 * Yeni bir 'User'i DB'ye kaydediyoruz.
+	 * @param createUserDto 
+	 * @returns 
+	 */
 	@Post()
-	create(@Body() createUserDto: CreateUserDto) {
-		return this.usersService.create(createUserDto);
+	async	create(@Body() createUserDto: CreateUserDto) {
+		const	newUser = this.usersService.create(createUserDto);
+		if (!newUser)
+			throw (new HttpException("@Post(): create(): User could not be created!",
+				HttpStatus.INTERNAL_SERVER_ERROR));
+		return (newUser);
 	}
 
+	/**
+	 * DB'den butun 'User' verilerini donduruyoruz.
+	 * @returns All Users.
+	 */
 	@Get()
 	findAll() {
 		return this.usersService.findAll();
 	}
 
+	/**
+	 * Verilen id'nin karsilik geldigi 'User' verisini donduruyoruz.
+	 * @param id Istenilen 'user'in 'id'si.
+	 * @returns 'user'.
+	 */
 	@Get(':id')
-	findOne(@Param('id') id: string) {
-		return this.usersService.findOne(+id);
+	async	findOne(@Param('id') id: string) {
+		// const	tmpUser = this.usersService.findOne(Number(id));
+		const	tmpUser = await this.usersService.findOne(+id);
+		if (!tmpUser)
+			throw (new NotFoundException("@Get(':id'): findOne(): id: User does not exist!"));
+		return (tmpUser);
 	}
 
+	/**
+	 * DB'de var olan User verisini guncellemek icin.
+	 * @param id 
+	 * @param updateUserDto 
+	 * @returns 
+	 */
 	@Patch(':id')
-	update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-		return this.usersService.update(+id, updateUserDto);
+	async	update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+		const	tmpUser = await this.usersService.update(+id, updateUserDto);
+		if (!tmpUser)
+			throw (new HttpException("@Patch(':id'): update(): id: User does not exist!",
+				HttpStatus.INTERNAL_SERVER_ERROR));
+		// if (updateUserDto !== tmpUser)
+		// if (!isEqual(updateUserDto, tmpUser)) // Buradaki function 'lodash' kutuphanesi icerisindeki objenin tipini ve icerisindeki degerlerini de esit mi diye kontrol ediyor.
+		// if (JSON.stringify(UpdateUserDto) !== JSON.stringify(tmpUser))
+			// throw (new HttpException("@Patch(':id'): update(): id: User's in DB but User does not update!",
+				// HttpStatus.INTERNAL_SERVER_ERROR));
+		return ({message: "User updated successfully."});
 	}
 
+	/**
+	 * Disaridan 'string' olarak aldigimiz 'id' parametremizi
+	 *  Number() ile 'number' tipine ceviriyoruz.
+	 * @param id String
+	 * @returns 
+	 */
 	@Delete(':id')
-	remove(@Param('id') id: string) {
+	async	remove(@Param('id') id: string) {
+		const	tmpUser = await this.usersService.findOne(Number(id));
+		if (!tmpUser)
+			throw (new NotFoundException("@Delete(':id'): remove(): id : User does not exist!")); // Eger 'user' olusturulamazsa exception atiyoruz.
 		return this.usersService.remove(+id);
 	}
 }
